requirements.txt
streamlit==1.32.2
requests==2.31.0
python-dotenv==1.0.1
watchdog==4.0.0
opea-comps==1.2 